Organize folder structure: Add folder pages
- Move to folder pages component that match a route and renders the
whole main section
- Renamed Main and Courses to contain Routes in their name
- Share NotFoundPage and rename Welcome to Home
